Carga de docentes y paginacion en caracterización docente, formulario con carga de datos reales, creación de nuevas cualificaciones x cada docente.

+++ PROYECTOS_CUALIFICACION/BACKEND/src/Models/cualificacion.js
@@ -8,7 +8,7 @@ const Cualificacion = sequelize.define('Cualificacion', {
     autoIncrement: true
   },
   id_docente: {
-    type: DataTypes.INTEGER, // O UUID si usaste UUID en Usuarios_Docentes
+    type: DataTypes.BIGINT, 
     allowNull: false
   },
   id_formacion: {
